refactor(deploy): move the workbench adapters into workbench-cli behind a factory
