+++ packages/@sanity/cli-core/src/_exports/deploy.ts
@@ -1,0 +1,35 @@
+export {
+  type DeployAdapter,
+  type DeployAppOptions,
+  type DeployedApplicationType,
+  type DeployedExpose,
+  type DeployFlags,
+  type DeploymentFile,
+  type DeploymentPlan,
+  type DeployResult,
+  type DeployState,
+  executeDeploy,
+  isDeployable,
+  listDeploymentFiles,
+  newPlan,
+  planDeploy,
+} from '../deploy/adapter.js'
+export {
+  APP_ID_NOT_FOUND_IN_ORGANIZATION,
+  type AppDeployTargetResolution,
+  checkOrganizationId,
+  checkProjectId,
+  type CommonDeployTargetResolution,
+  type DeployCheck,
+  type DeployTarget,
+  type DeployTargetApp,
+  type DeployTargetCoreApp,
+  describeAppTarget,
+  describeStudioTarget,
+  enforce,
+  getCoreAppUrl,
+  NO_ORGANIZATION_ID,
+  NO_PROJECT_ID,
+  type StudioDeployTargetResolution,
+  type TargetCheck,
+} from '../deploy/checks.js'
