+++ packages/@sanity/cli/src/actions/auth/ensureAuthenticated.ts
@@ -1,6 +1,7 @@
 import {
   type CLITelemetryStore,
   getCliToken,
+  getErrorMessage,
   type Output,
   type SanityOrgUser,
 } from '@sanity/cli-core'
@@ -8,7 +9,6 @@ import {isHttpError} from '@sanity/client'
 
 import {LoginError} from '../../errors/LoginError.js'
 import {getCliUser} from '../../services/user.js'
-import {getErrorMessage} from '../../util/getErrorMessage.js'
 import {login} from './login/login.js'
 
 /**
