+++ packages/@sanity/cli/src/actions/deploy/__tests__/urlUtils.test.ts
@@ -1,12 +1,6 @@
 import {describe, expect, test} from 'vitest'
 
-import {getCoreAppUrl, normalizeUrl, validateUrl} from '../urlUtils.js'
-
-describe('getCoreAppUrl', () => {
-  test('points at the org-scoped application route in the dashboard', () => {
-    expect(getCoreAppUrl('org-1', 'app-1')).toMatch(/\/@org-1\/application\/app-1$/)
-  })
-})
+import {normalizeUrl, validateUrl} from '../urlUtils.js'
 
 describe('validateUrl', () => {
   test('returns true for valid https URL', () => {
