+++ packages/@sanity/cli/src/actions/deploy/resolveDeployTarget.ts
@@ -1,4 +1,7 @@
-import {getApplication} from '@sanity/workbench-cli/deploy'
+import {
+  type AppDeployTargetResolution,
+  type StudioDeployTargetResolution,
+} from '@sanity/cli-core/deploy'
 
 import {
   getUserApplication,
@@ -6,50 +9,7 @@ import {
   type UserApplication,
   type UserApplicationResolved,
 } from '../../services/userApplications.js'
-import {APP_ID_NOT_FOUND_IN_ORGANIZATION} from '../../util/errorMessages.js'
 import {normalizeUrl, validateUrl} from './urlUtils.js'
-
-/** The application fields a deploy-target verdict carries for reporting. */
-interface DeployTargetApp {
-  appHost: string
-  id: string
-  title: string | null
-}
-
-/** A coreApp verdict also carries the organization, for the dashboard URL. */
-interface DeployTargetCoreApp extends DeployTargetApp {
-  organizationId: string
-}
-
-/**
- * The read-only outcome of resolving where a deploy would go.
- *
- * - `found` — the deploy targets this existing application
- * - `would-create` — nothing registered yet; a deploy would create it without prompting
- * - `needs-input` — config doesn't determine a target; a deploy would prompt
- *   (`existing` lists the applications a prompt would offer)
- * - `invalid` — the configured target can never resolve (bad host, unknown appId)
- * - `blocked` — resolution requires config that's missing (projectId/organizationId)
- *
- * Transport errors (network, permissions) throw — they're not verdicts.
- *
- * The user-applications resolvers carry the full {@link UserApplication} (the
- * real deploy needs it); the workbench resolvers and the report only need
- * {@link DeployTargetApp}, so `App` defaults to that widened shape.
- */
-type CommonDeployTargetResolution<App = DeployTargetApp> =
-  | {application: App; type: 'found'}
-  | {existing: App[]; type: 'needs-input'}
-  | {message: string; reason: 'app-not-found' | 'invalid-host'; type: 'invalid'}
-  | {message: string; type: 'blocked'}
-
-export type StudioDeployTargetResolution<App = DeployTargetApp> =
-  | CommonDeployTargetResolution<App>
-  | {appHost: string; type: 'would-create'}
-
-export type AppDeployTargetResolution<App = DeployTargetCoreApp> =
-  | CommonDeployTargetResolution<App>
-  | {type: 'would-create'}
 
 /**
  * Owns the studio deploy-target rules: the --url flag over studioHost config,
@@ -156,54 +116,6 @@ export async function resolveAppDeployTarget(options: {
   }
 
   return {type: 'would-create'}
-}
-
-/**
- * The dry-run counterpart to a workbench app's create-on-deploy: a configured
- * `appId` is looked up read-only, otherwise a coreApp would be created.
- * @internal
- */
-export async function resolveWorkbenchApp({
-  appId,
-}: {
-  appId: string | undefined
-}): Promise<AppDeployTargetResolution> {
-  return appId ? resolveAppById(appId) : {type: 'would-create'}
-}
-
-/**
- * The studio counterpart to {@link resolveWorkbenchApp}: a configured
- * `studioHost` would create that hostname, and without one a deploy would prompt.
- *
- * @internal
- */
-export async function resolveWorkbenchStudio({
-  appId,
-  studioHost,
-}: {
-  appId: string | undefined
-  studioHost: string | undefined
-}): Promise<StudioDeployTargetResolution> {
-  if (appId) return resolveAppById(appId)
-  if (studioHost) return {appHost: studioHost, type: 'would-create'}
-  return {existing: [], type: 'needs-input'}
-}
-
-async function resolveAppById(
-  appId: string,
-): Promise<CommonDeployTargetResolution<DeployTargetCoreApp>> {
-  const application = await getApplication(appId)
-  return application
-    ? {
-        application: {
-          appHost: application.slug ?? '',
-          id: application.id,
-          organizationId: application.organizationId,
-          title: application.title,
-        },
-        type: 'found',
-      }
-    : {message: APP_ID_NOT_FOUND_IN_ORGANIZATION, reason: 'app-not-found', type: 'invalid'}
 }
 
 async function listStudioApplications(
