+++ packages/@sanity/cli/src/actions/deploy/types.ts
@@ -1,17 +1,4 @@
-import {type CliConfig, type Output, type ProjectRootResult} from '@sanity/cli-core'
 import {z} from 'zod/mini'
-
-import {DeployCommand} from '../../commands/deploy.js'
-
-export type DeployFlags = DeployCommand['flags']
-
-export interface DeployAppOptions {
-  cliConfig: CliConfig
-  flags: DeployFlags
-  output: Output
-  projectRoot: ProjectRootResult
-  sourceDir: string
-}
 
 export const deployStudioSchemasAndManifestsWorkerData = z.object({
   configPath: z.string(),
