+++ packages/@sanity/workbench-cli/src/_exports/deploy.ts
@@ -1,19 +1,6 @@
 export {
-  type BrettInterface,
-  buildExposes,
-  type DeployedExpose,
-  summarizeExposes,
-} from '../actions/deploy/buildExposes.js'
+  createWorkbenchDeployAdapters,
+  type WorkbenchDeployHost,
+} from '../actions/deploy/adapters.js'
 export {checkBuiltOutput} from '../actions/deploy/checkBuiltOutput.js'
-export {
-  deployInstallationConfig,
-  resolveInstallationId,
-  summarizeInstallationConfig,
-} from '../actions/deploy/deployInstallationConfig.js'
-export {
-  type Application,
-  deployCoreApp,
-  deployStudio,
-  getApplication,
-} from '../actions/deploy/deployWorkbenchApp.js'
 export {getWorkbench} from '../actions/deploy/getWorkbench.js'
